Improvement of website design and functionality
Text in 'About me' section has more contrast. Button 'back' in the modal (SinglePost and Confirmation) was deleted. Clicking aside of the modal hides it. Added icons for 'delete' and 'edit' buttons. Added a defauld image when the image was not found. In 'Confirmation' window the 'delete' button is on left side and cancel on right.